ARM: fix at91x40 CPU ID
The old Atmel AT91x40 CPU family members do not have the ARM CP15 register,
and so do not have standard ARM processor IDs that can be read. Historically
we used the vendors chip ID as the CPU ID, and this is what we have set for
CONFIG_PROCESSOR_ID in the defaut kernel config (at91x40_defconfig).

The __get_cpu_architecture() code in arch/arm/kernel/setup.c though does a
more complete check on the value returned as the CPU ID, and its tests will
fail if we use the AT91x40 vendors chip ID.

So change the default configured processor ID to something more sensible
for this device. According to the ARM Architecture Reference Manual, "Main
ID register" section I think we want:

    BITS       VALUE
    31-24   --  0x41     (ARM Limited)
       23   --  0x1      (Architecture 4T)
    15-12   --  0x7      (Processor in ARM7 family)

Which gives a processor ID value of 0x41807000.

Signed-off-by: Greg Ungerer <gerg@uclinux.org>